+++ src/pyobsplot/obsplot.py
@@ -17,21 +17,31 @@ class Obsplot:
     with ObsplotJsdom depending on the renderer.
     """
 
-    def __new__(cls, renderer: str = "widget", debug: bool = False) -> Any:
+    def __new__(
+        cls, renderer: str = "widget", defaults: dict = {}, debug: bool = False
+    ) -> Any:
         """
         Main Obsplot class constructor. Returns a Creator instance depending on the
         renderer passed as argument.
+
+        Args:
+            renderer (str): renderer to be used.
+            defaults (dict): dict of default spec values.
+            debug (bool): if True, activate debug mode (for widget renderer only)
+
+        returns:
+            A Creator object of type depending of the renderer.
         """
 
         available_renderers = ["widget", "jsdom"]
 
         # Plot spec with the configured renderer
         if renderer == "widget":
-            return ObsplotWidgetCreator(debug)
+            return ObsplotWidgetCreator(defaults=defaults, debug=debug)
         elif renderer == "jsdom":
             if debug:
                 raise ValueError("debug mode is not available with jsdom renderer")
-            return ObsplotJsdomCreator()
+            return ObsplotJsdomCreator(defaults=defaults)
         else:
             raise ValueError(
                 f"""
@@ -46,17 +56,33 @@ class ObsplotCreator:
     Creator class.
     """
 
+    def __init__(self, defaults: dict = {}) -> None:
+        allowed_defaults = [
+            "marginTop",
+            "marginRight",
+            "marginBottom",
+            "marginLeft",
+            "margin",
+            "width",
+            "height",
+            "aspectRatio",
+            "style",
+        ]
+        for k in defaults:
+            if k not in allowed_defaults:
+                raise ValueError(
+                    f"{k} is not allowed in defaults.\nAllowed values: {allowed_defaults}."  # noqa: E501
+                )
+        self._defaults = defaults
+
     def get_spec(self, *args, **kwargs):
         """
         Extract plot specification from args and kwargs, taking into account
         the alternative specification syntaxes.
         """
 
-        # Only one arg which is already a widget output -> returns it as is
-        if len(args) == 1 and len(kwargs) == 0 and isinstance(args[0], (ObsplotWidget)):
-            spec = "widget"
         # Only one dict arg -> spec passed as dict
-        elif len(args) == 1 and len(kwargs) == 0 and isinstance(args[0], dict):
+        if len(args) == 1 and len(kwargs) == 0 and isinstance(args[0], dict):
             spec = args[0]
         # Only one kwarg called spec
         elif len(args) == 0 and len(kwargs) == 1 and "spec" in kwargs:
@@ -77,7 +103,8 @@ class ObsplotWidgetCreator(ObsplotCreator):
     Widget renderer Creator class.
     """
 
-    def __init__(self, debug: bool = False) -> None:
+    def __init__(self, defaults: dict = {}, debug: bool = False) -> None:
+        super().__init__(defaults)
         self._debug = debug
 
     def __call__(self, *args, **kwargs) -> ObsplotWidget:
@@ -85,9 +112,7 @@ class ObsplotWidgetCreator(ObsplotCreator):
         Method called whent an instance is called.
         """
         spec = self.get_spec(*args, **kwargs)
-        if spec == "widget":
-            return args[0]
-        return ObsplotWidget(spec, debug=self._debug)
+        return ObsplotWidget(spec, defaults=self._defaults, debug=self._debug)
 
 
 class ObsplotJsdomCreator(ObsplotCreator):
@@ -95,14 +120,12 @@ class ObsplotJsdomCreator(ObsplotCreator):
     Jsdom renderer Creator class.
     """
 
-    def __init__(self) -> None:
-        pass
+    def __init__(self, defaults: dict = {}) -> None:
+        super().__init__(defaults)
 
     def __call__(self, *args, **kwargs) -> None:
         """
         Method called whent an instance is called.
         """
         spec = self.get_spec(*args, **kwargs)
-        if spec == "widget":
-            raise ValueError("Incorrect plot specification")
-        display(ObsplotJsdom(spec).plot())
+        display(ObsplotJsdom(spec, defaults=self._defaults).plot())