+++ tests/python/test_parsing.py
@@ -8,12 +8,72 @@ import pandas as pd
 import polars as pl
 import datetime
 
-from pyobsplot import jsmodules
+from pyobsplot import jsmodules, Plot
 from pyobsplot.parsing import SpecParser, js
 from pyobsplot.data import pd_to_arrow, pl_to_arrow
 
 
 class TestSpecParser:
+    def test_specparser_init(self):
+        parser = SpecParser()
+        assert parser.renderer == "widget"
+        assert parser._spec == {}
+        assert parser._defaults == {}
+        parser = SpecParser(renderer="widget", defaults={"foo": "bar"})
+        assert parser.renderer == "widget"
+        assert parser._spec == {}
+        assert parser._defaults == {"foo": "bar"}
+        parser = SpecParser(renderer="jsdom")
+        assert parser.renderer == "jsdom"
+        assert parser._spec == {}
+        assert parser._defaults == {}
+        parser = SpecParser(renderer="jsdom", defaults={"foo": "bar"})
+        assert parser.renderer == "jsdom"
+        assert parser._spec == {}
+        assert parser._defaults == {"foo": "bar"}
+
+    def test_parse_plot_mark(self):
+        parser = SpecParser()
+        parser.spec = Plot.lineY([1, 2, 3])
+        assert parser.spec == {
+            "marks": [
+                {
+                    "pyobsplot-type": "function",
+                    "module": "Plot",
+                    "method": "lineY",
+                    "args": ([1, 2, 3],),
+                }
+            ]
+        }
+
+
+class TestMergeDefaults:
+    def test_merge_without_defaults(self):
+        spec = {"width": 100, "color": {"legend": True}}
+        parser = SpecParser(renderer="widget")
+        parser.spec = spec
+        parsed = parser.parse_spec()
+        assert parsed == spec
+        parser = SpecParser(renderer="jsdom")
+        parser.spec = spec
+        parsed = parser.parse_spec()
+        assert parsed == spec
+
+    def test_merge_with_defaults(self):
+        spec = {"width": 100, "color": {"legend": True}}
+        defaults = {"width": 200, "style": {"color": "red"}}
+        merged = {"width": 100, "color": {"legend": True}, "style": {"color": "red"}}
+        parser = SpecParser(renderer="widget", defaults=defaults)
+        parser.spec = spec
+        parsed = parser.parse_spec()
+        assert parsed == merged
+        parser = SpecParser(renderer="jsdom", defaults=defaults)
+        parser.spec = spec
+        parsed = parser.parse_spec()
+        assert parsed == merged
+
+
+class TestParse:
     def test_parse_none(self):
         assert SpecParser().parse(None) is None
 
@@ -232,58 +292,4 @@ class TestSpecParser:
         assert parser.serialize_data()[2] == {
             "pyobsplot-type": "DataFrame",
             "value": pd_to_arrow(df_pd),
-        }
-
-
-class TestJsModules:
-    def test_js_modules(self):
-        assert jsmodules.Plot.foo() == {
-            "pyobsplot-type": "function",
-            "module": "Plot",
-            "method": "foo",
-            "args": (),
-        }
-        assert jsmodules.d3.bar() == {
-            "pyobsplot-type": "function",
-            "module": "d3",
-            "method": "bar",
-            "args": (),
-        }
-        assert jsmodules.Math.baz() == {
-            "pyobsplot-type": "function",
-            "module": "Math",
-            "method": "baz",
-            "args": (),
-        }
-        assert jsmodules.Math.baz() == {
-            "pyobsplot-type": "function",
-            "module": "Math",
-            "method": "baz",
-            "args": (),
-        }
-
-    def test_js_modules_args(self):
-        assert jsmodules.Plot.foo(1, "bar") == {
-            "pyobsplot-type": "function",
-            "module": "Plot",
-            "method": "foo",
-            "args": (1, "bar"),
-        }
-        assert jsmodules.Plot.foo([1, 2], {"x": "foo"}) == {
-            "pyobsplot-type": "function",
-            "module": "Plot",
-            "method": "foo",
-            "args": ([1, 2], {"x": "foo"}),
-        }
-
-    def test_js_modules_kwargs(self):
-        with pytest.raises(ValueError, match="kwargs must not be passed to d3\\.foo.*"):
-            jsmodules.d3.foo(x=1)
-        with pytest.raises(ValueError, match="kwargs must not be passed to d3\\.bar.*"):
-            jsmodules.d3.bar(12, x=1)
-
-    def test_js(self):
-        assert js("d => d.foo") == {
-            "pyobsplot-type": "js",
-            "value": "d => d.foo",
         }