+++ NEWS.md
@@ -1,7 +1,8 @@
 ## pyobsplot (development version)
 
 - Add ability to specify some default spec values to plot generator objects.
-- range objects are correctly serialized as lists for `jsdom` renderer.
+- `range` objects are correctly serialized as lists for `jsdom` renderer.
+- Fix: don't add styles to svg or html output if styles are already present.
 
 
 ## pyobsplot 0.3.1